requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
python-dotenv
PyJWT
cryptography
twilio
firebase-admin